backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.20
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
chromadb>=0.5.0
pypdf==5.1.0
python-dotenv==1.0.1
pydantic==2.10.6
tiktoken==0.8.0
sentence-transformers>=2.2.0
